Move LoaderStats down to `/base/`
