+++ src/vs/workbench/contrib/performance/browser/perfviewEditor.ts
@@ -18,7 +18,7 @@ import * as perf from 'vs/base/common/performance';
 import { IDisposable, dispose } from 'vs/base/common/lifecycle';
 import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
 import { writeTransientState } from 'vs/workbench/contrib/codeEditor/browser/toggleWordWrap';
-import { mergeSort } from 'vs/base/common/arrays';
+import { LoaderStats } from 'vs/base/common/amd';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
@@ -279,100 +279,6 @@ class PerfModelContentProvider implements ITextModelContentProvider {
 	}
 }
 
-abstract class LoaderStats {
-	abstract get amdLoad(): (string | number)[][];
-	abstract get amdInvoke(): (string | number)[][];
-	abstract get nodeRequire(): (string | number)[][];
-	abstract get nodeEval(): (string | number)[][];
-	abstract get nodeRequireTotal(): number;
-
-
-	static get(): LoaderStats {
-
-
-		const amdLoadScript = new Map<string, number>();
-		const amdInvokeFactory = new Map<string, number>();
-		const nodeRequire = new Map<string, number>();
-		const nodeEval = new Map<string, number>();
-
-		function mark(map: Map<string, number>, stat: LoaderEvent) {
-			if (map.has(stat.detail)) {
-				// console.warn('BAD events, DOUBLE start', stat);
-				// map.delete(stat.detail);
-				return;
-			}
-			map.set(stat.detail, -stat.timestamp);
-		}
-
-		function diff(map: Map<string, number>, stat: LoaderEvent) {
-			let duration = map.get(stat.detail);
-			if (!duration) {
-				// console.warn('BAD events, end WITHOUT start', stat);
-				// map.delete(stat.detail);
-				return;
-			}
-			if (duration >= 0) {
-				// console.warn('BAD events, DOUBLE end', stat);
-				// map.delete(stat.detail);
-				return;
-			}
-			map.set(stat.detail, duration + stat.timestamp);
-		}
-
-		const stats = mergeSort(require.getStats().slice(0), (a, b) => a.timestamp - b.timestamp);
-
-		for (const stat of stats) {
-			switch (stat.type) {
-				case LoaderEventType.BeginLoadingScript:
-					mark(amdLoadScript, stat);
-					break;
-				case LoaderEventType.EndLoadingScriptOK:
-				case LoaderEventType.EndLoadingScriptError:
-					diff(amdLoadScript, stat);
-					break;
-
-				case LoaderEventType.BeginInvokeFactory:
-					mark(amdInvokeFactory, stat);
-					break;
-				case LoaderEventType.EndInvokeFactory:
-					diff(amdInvokeFactory, stat);
-					break;
-
-				case LoaderEventType.NodeBeginNativeRequire:
-					mark(nodeRequire, stat);
-					break;
-				case LoaderEventType.NodeEndNativeRequire:
-					diff(nodeRequire, stat);
-					break;
-
-				case LoaderEventType.NodeBeginEvaluatingScript:
-					mark(nodeEval, stat);
-					break;
-				case LoaderEventType.NodeEndEvaluatingScript:
-					diff(nodeEval, stat);
-					break;
-			}
-		}
-
-		let nodeRequireTotal = 0;
-		nodeRequire.forEach(value => nodeRequireTotal += value);
-
-		function to2dArray(map: Map<string, number>): (string | number)[][] {
-			let res: (string | number)[][] = [];
-			map.forEach((value, index) => res.push([index, value]));
-			return res;
-		}
-
-		return {
-			amdLoad: to2dArray(amdLoadScript),
-			amdInvoke: to2dArray(amdInvokeFactory),
-			nodeRequire: to2dArray(nodeRequire),
-			nodeEval: to2dArray(nodeEval),
-			nodeRequireTotal
-		};
-	}
-}
-
 class MarkdownBuilder {
 
 	value: string = '';
@@ -393,34 +299,6 @@ class MarkdownBuilder {
 	}
 
 	table(header: string[], rows: Array<Array<{ toString(): string } | undefined>>) {
-		let lengths: number[] = [];
-		header.forEach((cell, ci) => {
-			lengths[ci] = cell.length;
-		});
-		rows.forEach(row => {
-			row.forEach((cell, ci) => {
-				if (typeof cell === 'undefined') {
-					cell = row[ci] = '-';
-				}
-				const len = cell.toString().length;
-				lengths[ci] = Math.max(len, lengths[ci]);
-			});
-		});
-
-		// header
-		header.forEach((cell, ci) => { this.value += `| ${cell + ' '.repeat(lengths[ci] - cell.toString().length)} `; });
-		this.value += '|\n';
-		header.forEach((_cell, ci) => { this.value += `| ${'-'.repeat(lengths[ci])} `; });
-		this.value += '|\n';
-
-		// cells
-		rows.forEach(row => {
-			row.forEach((cell, ci) => {
-				if (typeof cell !== 'undefined') {
-					this.value += `| ${cell + ' '.repeat(lengths[ci] - cell.toString().length)} `;
-				}
-			});
-			this.value += '|\n';
-		});
+		this.value += LoaderStats.toMarkdownTable(header, rows);
 	}
 }
